mmu 32bit

+++ src/cpu.h
@@ -13,8 +13,8 @@ typedef struct instruction_tag instruction_t;
 #endif
 
 typedef void * context_t;
-typedef uint8_t *(*mmu_v2r_t)(context_t ctx, uint16_t vaddr);
-typedef uint16_t (*mmu_r2v_t)(context_t ctx, uint8_t *raddr);
+typedef uint8_t *(*mmu_v2r_t)(context_t ctx, uint32_t vaddr);
+typedef uint32_t (*mmu_r2v_t)(context_t ctx, uint8_t *raddr);
 typedef void (*syscall_t)(context_t ctx);
 typedef void (*syscall_string_t)(context_t ctx, char *str, size_t size, uint8_t id);
